+++ src/main/java/tornadofx/Component.kt
@@ -1137,9 +1137,27 @@ fun <U : UIComponent> U.whenDocked(listener: (U) -> Unit) {
 }
 
 @Suppress("UNCHECKED_CAST")
+fun <U : UIComponent> U.whenDockedOnce(listener: (U) -> Unit) {
+    if (onDockListeners == null) onDockListeners = mutableListOf()
+    onDockListeners!!.add {
+        onDockListeners!!.remove(listener)
+        listener(this)
+    }
+}
+
+@Suppress("UNCHECKED_CAST")
 fun <U : UIComponent> U.whenUndocked(listener: (U) -> Unit) {
     if (onUndockListeners == null) onUndockListeners = mutableListOf()
     onUndockListeners!!.add(listener as (UIComponent) -> Unit)
+}
+
+@Suppress("UNCHECKED_CAST")
+fun <U : UIComponent> U.whenUndockedOnce(listener: (U) -> Unit) {
+    if (onUndockListeners == null) onUndockListeners = mutableListOf()
+    onUndockListeners!!.add {
+        onUndockListeners!!.remove(listener)
+        listener(this)
+    }
 }
 
 abstract class Fragment @JvmOverloads constructor(title: String? = null, icon: Node? = null) : UIComponent(title, icon)